requirements.txt
# Core dependencies
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Web UI
streamlit>=1.30.0

# CLI utilities
rich>=13.0.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
